chore(docs): Optimized array indexing in the command bindings example
# Project State

## Current Focus
Optimized array indexing in the command bindings example

## Context
The change improves performance in the command bindings example by removing unnecessary casting operations in the array indexing logic.

## Completed
- [x] Removed redundant `as usize` cast in array index calculation
- [x] Simplified index calculation for better readability and performance

## In Progress
- [x] None - this is a completed optimization

## Blockers
- None

## Next Steps
1. Verify no regression in command bindings behavior
2. Consider similar optimizations in other examples if needed
---
category: chore
scope: docs
---

+++ examples/_cookbook/command_bindings.rs
@@ -333,7 +333,7 @@ impl Widget for CommandBindings {
         let auto_str = if self.paused { "⏸ PAUSED" } else { "▶ RUNNING" };
         let status = format!("  {}  |  tick: {}  |  s=refresh p=pause  |  q=quit", auto_str, self.tick);
         for (i, c) in status.chars().enumerate().take(w - 2) {
-            let idx = ((h - 1) * w + 1 + i) as usize;
+            let idx = (h - 1) * w + 1 + i;
             if idx < p.cells.len() {
                 p.cells[idx] = Cell { char: c, fg: t.fg_muted, bg: t.surface_elevated, style: Styles::empty(), transparent: false, skip: false };
             }
